teasing out multistage components into separate module
